+++ src/router/AppRouter.tsx
@@ -9,6 +9,7 @@ import PublicarPedido from '../pages/comprador/PublicarPedido';
 import MisCotizacionesComprador from '../pages/comprador/MisCotizacionesComprador';
 import MisOrdenesComprador from '../pages/comprador/MisOrdenesComprador';
 import ChatComprador from '../pages/comprador/ChatComprador';
+import DetallePedidoComprador from '../pages/comprador/DetallePedidoComprador';
 
 import DashboardProveedor from '../pages/proveedor/DashboardProveedor';
 import PedidosDisponibles from '../pages/proveedor/PedidosDisponibles';
@@ -52,6 +53,7 @@ export function AppRouter() {
           <Route path="/comprador/cotizaciones" element={<MisCotizacionesComprador />} />
           <Route path="/comprador/ordenes" element={<MisOrdenesComprador />} />
           <Route path="/comprador/chat" element={<ChatComprador />} />
+          <Route path="/comprador/pedidos/:id" element={<DetallePedidoComprador />} />
 
           <Route path="/proveedor" element={<DashboardProveedor />} />
           <Route path="/proveedor/pedidos" element={<PedidosDisponibles />} />